Moving sanitizers from build/ to base/
We have to move the content of the 'webrtc/build/' directory up of one level (see crbug.com/611808 for further details).

To avoid a collision with the DEPSed build/ at the top level we are moving all the content of 'webrtc/build/' to different locations (see webrtc:7030 for the suggested locations).

BUG=webrtc:7030
NOTRY=True

Review-Url: https://codereview.webrtc.org/2651913002
Cr-Original-Commit-Position: refs/heads/master@{#16259}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 18e83ea598ae8ed0446169d35a34cc5d14dc71a0