Fix flaky ViEEncoder unit test.
The flaky test was introduced in ad9010c9836, and is essentially a race
where the ViE Encoder has already configured the quality scaler on the
encoder thread before we've updated the ScalingSettings. This CL adds
a forced reconfiguration of the quality scaler to avoid this issue.

BUG=None
TBR=sprang@webrtc.org

Review-Url: https://codereview.webrtc.org/2695873004
Cr-Original-Commit-Position: refs/heads/master@{#16612}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: b83797bd7a7239049dcdd8d39a304769c3270a41