TrackMediaInfoMap: Allow same SSRC for send and receive side.
Running video loopback on https://appr.tc/ revealed that it is possible
to use the same SSRC for a local and remote audio or video track. This
caused a DCHECK crash. The constructor of TrackMediaInfoMap is updated
to support this mapping and the unittest is updated (moved and modified
a test from being a death test to being a non-death test).

I've verified that this fixes the bug.

BUG=chromium:693087

Review-Url: https://codereview.webrtc.org/2703783002
Cr-Original-Commit-Position: refs/heads/master@{#16713}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: fe90ad195f16e654e0b6281641b08d8821236308