Add a PrintTo function for rtc::Optional to aid with testing.
gtest can print objects if they have an operator<< or a PrintTo
function in the same namespace as the object's class. Since
std::optional does not seem to have an operator<<, it'd be preferable
not to rely on rtc::Optional being printable through operator<<.

Currently, gtest errors will just dump the raw bytes of
rtc::Optionals, which make them really annoying to work with in tests.

BUG=webrtc:7196

Review-Url: https://codereview.webrtc.org/2704483002
Cr-Original-Commit-Position: refs/heads/master@{#16717}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: e5c27a5db6dab92176f90c4a0cd5a77c6d166a58