Fix partial availability warnings on Mac AppRTCMobile
The partial availability problem aries from the	fact that the minimum
supported OSX version is set to 10.9, but AppRTCMobile is using
functions available only in 10.10 and later. The minimum OSX version is
set as a declare_args() in build/config/mac/mac_sdk.gni, which makes it
difficult to override for just the AppRTCMobile target in WebRTC.

Instead, this CL solves the problem for now by removing the usage of the
10.10 function, which is trivial.

Also, the flag:
'extra_substitutions = [ "MACOSX_DEPLOYMENT_TARGET=10.8" ]'
is removed since it has no effect.

BUG=webrtc:4695

Review-Url: https://codereview.webrtc.org/2710493002
Cr-Original-Commit-Position: refs/heads/master@{#16726}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: a445b9bca72606e55970135f0f1023e7c486d31e