Implement operator<< for AudioCodec
It's annoying to have to re-implement this every time I need a debug
printout.

Declared inline, so that there'll be zero runtime overhead.

This CL also modifies a unit test so that it will make use of the new
operator<< in case it finds errors.

BUG=none

Review-Url: https://codereview.webrtc.org/2705203002
Cr-Original-Commit-Position: refs/heads/master@{#16749}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: b94491d7904feb9b072b1adaa9454c9615fba5f8